move outliers_influence tests to test_influence
